rename parse scripts
